fix #701: move to cache package
